Split ViconJson into ViconJsonZmq and ViconJsonFile
The recent change broke the "load file as fallback" mode.  This this
anyway seems to be meant only for testing, better split it into two
separate classes to keep the code more simple.
This also makes it more explicit which mode is used (thus avoids
accidentally ending up with something unexpected).