Add JsonReceiver for testing
Replacement for ViconReceiver which loads a single frame from a JSON
file and simply always returns that frame.  Equivalent to the
corresponding Python implementation.

New data format version 3, which matches with how cereal serialises
ViconFrame to JSON.  Unfortunately, it does not support converting the
std::map to/from a JSON dictionary but instead produces a list of `{key:
.., value: ..}` dictionaries.