test: Change working directory of Python tests
This is needed as by default they are run in the package's source root
directory, which results in `import vicon_transformer` to prefer the
local package instead of the one from the install space.  However, the
version from the install space is needed as only there the pybind11
module is available.