Implement o80 driver and standalone
Unfortunately the shared memory used by o80 requires fixed-size data
structures, so using `ViconFrame` is not possible due to the dynamically
allocated subjects map.
Instead add `FixedSizeViconFrame` which requires the number of subjects
as template argument and does not include the subject names (as
`std::string` is also problematic).
To still know which subject data belongs to which actual subject, a
function `map_name_to_index` is given to the driver as template
argument.  This function is expected to map the subjects name to an
index in the subjects array.
The client code then needs access to the same mapping information to
access the data of the desired subjects.

As the number and names of subjects is application-specific, this part
of the implementation is moved to a separate header `pam_vicon_o80.hpp`
with the intention to move it to a separate package soon.