Record synchronised tennicam-vicon-trajectory
Add new script `record_tennicam_vicon_trajectory.py` that connects to
both tennicam_client and pam_vicon to record the trajectory of a ball
mounted on the LED stick with both systems (tennicam records the ball
positions while Vicon records the LED stick).

Assuming that the ball is mounted at the origin of the LED stick, the
resulting trajectory can be used to compute the spatial transformation
between the two systems.