Catch errors during recording
`tennicam_frontend.read()` raises an exception if SIGINT is received
while waiting for the observation.  This is expected and not an actual
error, so catch it and silently break the loop.
Only if an exception is raised but no sigint is reported, something is
wrong. Print the error message in this case.